+++ src/main/java/br/com/sgfly/model/DadosReceita.java
@@ -3,6 +3,7 @@ package br.com.sgfly.model;
 import br.com.sgfly.model.entities.Receita;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -10,7 +11,7 @@ public record DadosReceita(
         Long id,
         String descricao,
         Boolean fixo,
-        Integer valor,
+        BigDecimal valor,
         DadosPlanoContas planoContas,
         String observacao,
         @JsonFormat(pattern = "dd/MM/yyyy")
